+++ backend/src/main/java/com/ssafy/togeball/domain/chatroom/repository/RecruitTagRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RecruitTagRepository extends JpaRepository<RecruitTag, Long> {
+public interface RecruitTagRepository extends JpaRepository<RecruitTag, Integer> {
 
-    Optional<RecruitTag> findByRecruitChatroomIdAndTagId(Long chatroomId, Long tagId);
+    Optional<RecruitTag> findByRecruitChatroomIdAndTagId(Integer chatroomId, Integer tagId);
 }
