+++ backend/src/main/java/com/ssafy/togeball/domain/tag/repository/UserTagRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserTagRepository extends JpaRepository<UserTag, Long> {
+public interface UserTagRepository extends JpaRepository<UserTag, Integer> {
 
-    Optional<UserTag> findByTagIdAndUserId(Long tagId, Long userId);
+    Optional<UserTag> findByTagIdAndUserId(Integer tagId, Integer userId);
 
-    List<UserTag> findByUserId(Long userId);
+    List<UserTag> findByUserId(Integer userId);
 
-    List<UserTag> findByTagId(Long tagId);
+    List<UserTag> findByTagId(Integer tagId);
 }
